Adicionando exceptions

+++ src/main/java/com/medcontrol/medcontrol/service/MedicamentoService.java
@@ -1,5 +1,6 @@
 package com.medcontrol.medcontrol.service;
 
+import com.medcontrol.medcontrol.exception.MedicamentoNotFoundException;
 import com.medcontrol.medcontrol.model.MedicamentoModel;
 import com.medcontrol.medcontrol.repository.MedicamentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +22,8 @@ public class MedicamentoService {
     }
 
     public MedicamentoModel getMedicamentoById(Long id) {
-        return medicamentoRepository.findById(id).orElse(null);
+        return medicamentoRepository.findById(id)
+                .orElseThrow(() -> new MedicamentoNotFoundException("Medicamento não encontrado com o ID: " + id));
     }
 
     public MedicamentoModel createMedicamento(MedicamentoModel medicamento) {
@@ -29,6 +31,11 @@ public class MedicamentoService {
     }
 
     public void deleteMedicamento(Long id) {
-        medicamentoRepository.deleteById(id);
+        if (!medicamentoRepository.existsById(id)) {
+            throw new MedicamentoNotFoundException("O id para este nmedicamento não existe ou já foi excluído");
+        } else {
+            medicamentoRepository.deleteById(id);
+        }
+
     }
 }
